Yay! A blank screen!
Seriously, though. It's not exactly  amazing, but it's a start. Native
LWJGL bindings  for Windows, Mac  OS X, and  Linux have been  added. A
Maven repository  hosted somewhere on  Google Code for LWJGL  has also
been referenced. There are a lot of scary, large binary files in ext/,
but I doubt they'll really change much, so it's not much of an issue.